requirements.txt
streamlit>=1.38.0
ultralytics
pymupdf
Pillow
matplotlib
numpy>=2.1.0
